Moves useful code to central places
Moves some of the testing utility functions from test_validate.c into validate_testing.c and validate_testing.h

Also, moves some of the useful macros in validate.c into jvst_macros.h
