+++ src/validate.c
@@ -10,6 +10,8 @@
 
 #include <assert.h>
 
+#include "jvst_macros.h"
+
 #include "sjp_parser.h"
 #include "sjp_testing.h"
 #include "ast.h"
@@ -23,21 +25,6 @@
    } while(0)
 #else
 #  define SHOULD_NOT_REACH(v) return INVALID(v, "internal error: SHOULD_NOT_REACH()");
-#endif
-
-#define NCOUNTERS 2
-
-#define ARRAYLEN(x) (sizeof (x) / sizeof (x)[0])
-#define STRINGIFY0(x) # x
-#define STRINGIFY(x) STRINGIFY0(x)
-#define SYMCAT0(x,y) x ## y
-#define SYMCAT(x,y) SYMCAT0(x,y)
-#define SYMCAT3(x,y,z) SYMCAT(SYMCAT(x,y),z)
-
-#ifdef static_assert
-#  define STATIC_ASSERT(x,name) static_assert((x),STRINGIFY(name))
-#else
-#  define STATIC_ASSERT(x,name) struct { char tmp[2*(x)-1]; } SYMCAT3(static_assert_, name, __LINE__)
 #endif
 
 #define INVALID(v, ...) invalid((v), __FILE__, __LINE__, __VA_ARGS__)
